backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "ProseProof"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
]